+++ src/main/java/com/bbzavrsni/zavrsni/model/pojo/ApplicationUser.java
@@ -12,7 +12,7 @@ public class ApplicationUser implements Serializable {
     @Serial
     private static final long serialVersionUID = 8533039291044343363L;
 
-    private Integer UID;
+    private Long UID;
     private String username;
     private String password;
     private List<SimpleGrantedAuthority> authorities;
@@ -41,11 +41,11 @@ public class ApplicationUser implements Serializable {
         this.authorities = authorities;
     }
 
-    public Integer getUID() {
+    public Long getUID() {
         return UID;
     }
 
-    public void setUID(Integer UID) {
+    public void setUID(Long UID) {
         this.UID = UID;
     }
 }
